+++ src/main/java/com/kakaoasset/portfolio/controller/SectorController.java
@@ -1,6 +1,6 @@
-package com.kakaoasset.portfolio.elasticsearch.elasticAPI.controller;
+package com.kakaoasset.portfolio.controller;
 
-import com.kakaoasset.portfolio.elasticsearch.elasticAPI.service.SectorService;
+import com.kakaoasset.portfolio.service.SectorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
